core/animation: widen Sprite fade rescale multiply to avoid signed overflow (finding 1)

+++ core/animation.h
@@ -1204,7 +1204,8 @@ public:
     // only; indefinite ones skip fade-out).
     int fade_total = this->fade_in_duration + this->fade_out_duration;
     if (duration >= 0 && fade_total > duration) {
-      this->fade_in_duration = duration * this->fade_in_duration / fade_total;
+      this->fade_in_duration =
+          static_cast<long long>(duration) * this->fade_in_duration / fade_total;
       this->fade_out_duration = duration - this->fade_in_duration;
     }
   }
